refactor(core): use trait setlocation type field
